Update FireBall.java
I suspect the crashing error is that the Fireball's Box2D body is destroyed twice. First at any point after destroyTimer > 3, where during that time after already being flagged to destroy it collides with another entity flagging it a second time. Box2D updates the world physics first in the levelscreen update method. If during the intermediary step between a world step update, the Box2D body is flagged to destroy twice. The first flag will delete the body, then attempt to delete it again either in this inermediary step or in the next world step update.

If the committed solution doesn't work try adding the bodies to destroy to an array after contact, passing it to the levelscreen class and destroying the bodies BEFORE the world.step. Keep in mind the same issue of double deletion applies.